+++ client/config/config_test.go
@@ -17,16 +17,27 @@ import (
 )
 
 const (
+	chainID   = "test-chain"
 	nodeEnv   = "NODE"
 	testNode1 = "http://localhost:1"
 	testNode2 = "http://localhost:2"
 )
 
-// initClientContext initiates client Context for tests
+// initClientContext initiates client.Context for tests
 func initClientContext(t *testing.T, envVar string) (client.Context, func()) {
 	t.Helper()
+
+	clientCtx, cleanup, err := initClientContextWithTemplate(t, envVar, "", nil)
+	require.NoError(t, err)
+	require.Equal(t, chainID, clientCtx.ChainID)
+
+	return clientCtx, cleanup
+}
+
+// initClientContextWithTemplate initiates client.Context with custom config and template for tests
+func initClientContextWithTemplate(t *testing.T, envVar, customTemplate string, customConfig interface{}) (client.Context, func(), error) {
+	t.Helper()
 	home := t.TempDir()
-	chainID := "test-chain"
 	clientCtx := client.Context{}.
 		WithHomeDir(home).
 		WithViper("").
@@ -38,11 +49,89 @@ func initClientContext(t *testing.T, envVar string) (client.Context, func()) {
 		require.NoError(t, os.Setenv(nodeEnv, envVar))
 	}
 
-	clientCtx, err := config.ReadFromClientConfig(clientCtx)
-	require.NoError(t, err)
-	require.Equal(t, clientCtx.ChainID, chainID)
+	clientCtx, err := config.CreateClientConfig(clientCtx, customTemplate, customConfig)
+	return clientCtx, func() { _ = os.RemoveAll(home) }, err
+}
 
-	return clientCtx, func() { _ = os.RemoveAll(home) }
+func TestCustomTemplateAndConfig(t *testing.T) {
+	type GasConfig struct {
+		GasAdjustment float64 `mapstructure:"gas-adjustment"`
+	}
+
+	type CustomClientConfig struct {
+		config.Config `mapstructure:",squash"`
+
+		GasConfig GasConfig `mapstructure:"gas"`
+
+		Note string `mapstructure:"note"`
+	}
+
+	clientCfg := config.DefaultConfig()
+	// Overwrite the default keyring backend.
+	clientCfg.KeyringBackend = "test"
+
+	customClientConfig := CustomClientConfig{
+		Config: *clientCfg,
+		GasConfig: GasConfig{
+			GasAdjustment: 1.5,
+		},
+		Note: "Sent from the CLI.",
+	}
+
+	customClientConfigTemplate := config.DefaultClientConfigTemplate + `
+# This is the gas adjustment factor used by the tx commands.
+# Sets the default and can be overwriten by the --gas-adjustment flag in tx commands.
+gas-adjustment = {{ .GasConfig.GasAdjustment }}
+# Memo to include in all transactions.
+note = "{{ .Note }}"
+`
+
+	t.Run("custom template and config provided", func(t *testing.T) {
+		clientCtx, cleanup, err := initClientContextWithTemplate(t, "", customClientConfigTemplate, customClientConfig)
+		defer func() {
+			cleanup()
+			_ = os.Unsetenv(nodeEnv)
+		}()
+
+		require.NoError(t, err)
+		require.Equal(t, customClientConfig.KeyringBackend, clientCtx.Viper.Get(flags.FlagKeyringBackend))
+		require.Equal(t, customClientConfig.GasConfig.GasAdjustment, clientCtx.Viper.GetFloat64(flags.FlagGasAdjustment))
+		require.Equal(t, customClientConfig.Note, clientCtx.Viper.GetString(flags.FlagNote))
+	})
+
+	t.Run("no template and custom config provided", func(t *testing.T) {
+		_, cleanup, err := initClientContextWithTemplate(t, "", "", customClientConfig)
+		defer func() {
+			cleanup()
+			_ = os.Unsetenv(nodeEnv)
+		}()
+
+		require.Error(t, err)
+	})
+
+	t.Run("default template and custom config provided", func(t *testing.T) {
+		clientCtx, cleanup, err := initClientContextWithTemplate(t, "", config.DefaultClientConfigTemplate, customClientConfig)
+		defer func() {
+			cleanup()
+			_ = os.Unsetenv(nodeEnv)
+		}()
+
+		require.NoError(t, err)
+		require.Equal(t, customClientConfig.KeyringBackend, clientCtx.Viper.Get(flags.FlagKeyringBackend))
+		require.Nil(t, clientCtx.Viper.Get(flags.FlagGasAdjustment)) // nil because we do not read the flags
+	})
+
+	t.Run("no template and no config provided", func(t *testing.T) {
+		clientCtx, cleanup, err := initClientContextWithTemplate(t, "", "", nil)
+		defer func() {
+			cleanup()
+			_ = os.Unsetenv(nodeEnv)
+		}()
+
+		require.NoError(t, err)
+		require.Equal(t, config.DefaultConfig().KeyringBackend, clientCtx.Viper.Get(flags.FlagKeyringBackend))
+		require.Nil(t, clientCtx.Viper.Get(flags.FlagGasAdjustment)) // nil because we do not read the flags
+	})
 }
 
 func TestConfigCmdEnvFlag(t *testing.T) {
@@ -79,6 +168,7 @@ func TestConfigCmdEnvFlag(t *testing.T) {
 				cleanup()
 				_ = os.Unsetenv(nodeEnv)
 			}()
+
 			/*
 				env var is set with a flag