+++ testutil/network/network.go
@@ -588,7 +588,10 @@ func New(l Logger, baseDir string, cfg Config) (*Network, error) {
 		if err != nil {
 			return nil, err
 		}
-		srvconfig.WriteConfigFile(filepath.Join(nodeDir, "config", "app.toml"), appCfg)
+		err = srvconfig.WriteConfigFile(filepath.Join(nodeDir, "config", "app.toml"), appCfg)
+		if err != nil {
+			return nil, err
+		}
 
 		clientCtx := client.Context{}.
 			WithKeyringDir(clientDir).